+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - VIMEET-791 - Display conditional objective filter
+- VIMEET-1874 - Can participant scan in Product option
+- VIMEET-1972 - Display scan and contacts buttons by Product option 
+- VIMEET-1871 - As an admin, add promotion code to 
 
 ## [2.6.0] - 2019-05-16
 ### Added